+++ changelog.md
@@ -1,6 +1,6 @@
 # UnderScript Changelog
 
-## Version 0.47.0-1 (2021-12-12)
+## Version 0.47.0 (2021-12-12)
 ### New Features
 1. New display for underscript settings
 1. Added audio settings
@@ -9,6 +9,7 @@
 1. Expose more `events` functions
 1. Fix `user.ignore()`
 1. Added `setting.show()` to open the settings window
+1. Removed deprecated `underscript.addStyle()`
 
 ## Version 0.46.4 (2021-10-31)
 1. Fixed streamer mode chat